refactor migrations

+++ database/migrations/2022_08_23_162101_create_course_students_table.php
@@ -14,9 +14,9 @@ return new class extends Migration
     public function up()
     {
         Schema::create('course_students', function (Blueprint $table) {
-            $table->unsignedInteger('course_id');
+            $table->unsignedBigInteger('course_id');
             $table->foreign('course_id')->references('id')->on('courses');
-            $table->unsignedInteger('student_id');
+            $table->unsignedBigInteger('student_id');
             $table->foreign('student_id')->references('id')->on('students');
         });
     }
